Add files via upload
"Snake Water Gun" is a simple yet engaging game that is played between the user and the computer. The game offers three options - "snake", "water", and "gun". The user selects one option, and the computer randomly selects one as well. The game is then played according to a set of rules that determine which option wins over the other. The winner of the game is the one who wins the maximum number of rounds. This game is a great way to pass the time and can be enjoyed by players of all ages.